# packages/spec/swagger.json
{"openapi":"3.0.0","paths":{"/info":{"get":{"operationId":"getInfo","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AppInfoViewModel"}}}}},"tags":["App"]}},"/users":{"post":{"operationId":"createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateViewModel"}}}},"responses":{"201":{"description":""}},"tags":["users"]}},"/users/profile":{"post":{"operationId":"getUserProfile","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserViewModel"}}}}},"tags":["users"]}},"/users/{userId}/settings":{"put":{"operationId":"updateUserSettings","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSettingsViewModel"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserSettingsViewModel"}}}}},"tags":["users"]}},"/audits/list":{"post":{"operationId":"listAudits","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditListViewModel"}}}}},"tags":["audits"]}},"/jobs/history":{"get":{"operationId":"getJobsHistory","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditListViewModel"}}}}},"tags":["jobs"]}},"/jobs/applications/list":{"post":{"operationId":"listApplications","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiPaginatedRequest"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationListViewModel"}}}}},"tags":["applications"]}},"/jobs/applications/upload":{"post":{"operationId":"uploadApplications","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/ApplicationsUploadViewModel"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationViewModel"}}}}},"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"type":"object"}}}}}},"tags":["applications"]}},"/jobs/applications/bulk":{"post":{"operationId":"createApplications","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationViewModel"}}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["applications"]}},"/jobs/applications":{"post":{"operationId":"createApplication","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationViewModel"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["applications"]}},"/jobs/applications/{applicationId}":{"put":{"operationId":"updateApplication","parameters":[{"name":"applicationId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationViewModel"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationViewModel"}}}}},"tags":["applications"]},"get":{"operationId":"getApplication","parameters":[{"name":"applicationId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplicationViewModel"}}}}},"tags":["applications"]},"delete":{"operationId":"deleteApplication","parameters":[{"name":"applicationId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["applications"]}},"/jobs/companies":{"get":{"operationId":"getCompanies","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyListViewModel"}}}}},"tags":["companies"]}},"/jobs/companies/list":{"post":{"operationId":"getCompaniesList","parameters":[],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompanyFullListViewModel"}}}}},"tags":["companies"]}},"/jobs/companies/{companyId}":{"delete":{"operationId":"deleteCompany","parameters":[{"name":"companyId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["companies"]}},"/exercises/history":{"get":{"operationId":"getExercisesHistory","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuditListViewModel"}}}}},"tags":["exercises"]}},"/exercises/activities/list":{"post":{"operationId":"listActivities","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivityListViewModel"}}}}},"tags":["activities"]}},"/exercises/activities/strava/import":{"post":{"operationId":"importStravaActivities","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/ExerciseActivityUpload"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ActivityCreateViewModel"}}}}}},"tags":["activities"]}},"/exercises/activities/strava/upload":{"post":{"operationId":"uploadStravaActivities","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ActivityCreateViewModel"}}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["activities"]}},"/exercises/activities/strava/sync":{"post":{"operationId":"syncStravaActivities","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StravaTokenViewModel"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StravaTokenViewModel"}}}}},"tags":["activities"]}},"/exercises/activities/activity-types":{"get":{"operationId":"getActivityTypes","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ActivityTypeViewModel"}}}}}},"tags":["activities"]}},"/exercises/activities":{"post":{"operationId":"createActivity","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivityCreateViewModel"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivityViewModel"}}}}},"tags":["activities"]}},"/exercises/activities/{activityId}":{"get":{"operationId":"getActivity","parameters":[{"name":"activityId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivityViewModel"}}}}},"tags":["activities"]},"put":{"operationId":"updateActivity","parameters":[{"name":"activityId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivityViewModel"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivityViewModel"}}}}},"tags":["activities"]},"delete":{"operationId":"deleteActivity","parameters":[{"name":"activityId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":""}},"tags":["activities"]}},"/attribute-types":{"get":{"operationId":"getAttributeTypes","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AttributeTypeViewModel"}}}}}},"tags":["attributeTypes"]}}},"info":{"title":"API","description":"The main API for the UI","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"AppInfoViewModel":{"type":"object","properties":{"version":{"type":"string"}},"required":["version"]},"UserCreateViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"nickname":{"type":"string"}},"required":["firstName","email"]},"UserSettingsExerciseViewModel":{"type":"object","properties":{"weight":{"type":"number"}}},"UserSettingsViewModel":{"type":"object","properties":{"exercises":{"$ref":"#/components/schemas/UserSettingsExerciseViewModel"}},"required":["exercises"]},"UserViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"nickname":{"type":"string"},"id":{"type":"string"},"lastAccessed":{"type":"number"},"settings":{"$ref":"#/components/schemas/UserSettingsViewModel"}},"required":["firstName","email","id","lastAccessed","settings"]},"EnumTableNames":{"type":"number","enum":[1,2,3,4,5,6,7,8,9,10],"x-enum-varnames":["audits","exerciseActivities","exerciseActivityTypes","exerciseActivityAttributes","attributeTypes","jobApplications","jobCompanies","jobComments","unitTypes","users"]},"EnumAuditActionTypes":{"type":"string","enum":["Created","Updated","Deleted"],"x-enum-varnames":["created","updated","deleted"]},"EnumFeatures":{"type":"number","enum":[1,2,3],"x-enum-varnames":["jobs","exercises","system"]},"AuditDiffViewModel":{"type":"object","properties":{"action":{"allOf":[{"$ref":"#/components/schemas/EnumAuditActionTypes"}]},"field":{"type":"string"},"valueCurrent":{"type":"object"},"valuePrevious":{"type":"object"}},"required":["action","field"]},"AuditViewModel":{"type":"object","properties":{"entity":{"allOf":[{"$ref":"#/components/schemas/EnumTableNames"}]},"action":{"allOf":[{"$ref":"#/components/schemas/EnumAuditActionTypes"}]},"feature":{"allOf":[{"$ref":"#/components/schemas/EnumFeatures"}]},"id":{"type":"string"},"userId":{"type":"string"},"entityId":{"type":"string"},"diff":{"type":"array","items":{"$ref":"#/components/schemas/AuditDiffViewModel"}},"dateCreated":{"type":"number"}},"required":["entity","action","feature","id","userId","entityId","diff","dateCreated"]},"AuditListViewModel":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/AuditViewModel"}},"total":{"type":"number"}},"required":["data"]},"ApiPaginatedRequest":{"type":"object","properties":{"start":{"type":"number"},"limit":{"type":"number"},"page":{"type":"number"}},"required":["start","limit","page"]},"EnumApplicationStatus":{"type":"number","enum":[-1,0,1,2,3,4,5,6,7],"x-enum-varnames":["Applied","CurrentWeek","Initial","Interviewing","InterviewedAndRejected","Rejected","Declined","Accepted","Ghosted"]},"EnumLocationTypes":{"type":"number","enum":[1,2,3],"x-enum-varnames":["Remote","Hybrid","OnSite"]},"CompanyViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"id":{"type":"string"},"name":{"type":"string"}},"required":["id","name"]},"CommentViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"id":{"type":"string"},"comment":{"type":"string"},"applicationId":{"type":"string"}},"required":["id","comment","applicationId"]},"EnumLinkType":{"type":"number","enum":[1,2],"x-enum-varnames":["To","From"]},"ApplicationLinkViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"status":{"allOf":[{"$ref":"#/components/schemas/EnumApplicationStatus"}]},"type":{"allOf":[{"$ref":"#/components/schemas/EnumLinkType"}]},"id":{"type":"string"},"positionTitle":{"type":"string"}},"required":["type","id"]},"ApplicationViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"status":{"allOf":[{"$ref":"#/components/schemas/EnumApplicationStatus"}]},"locationType":{"allOf":[{"$ref":"#/components/schemas/EnumLocationTypes"}]},"id":{"type":"string"},"site":{"type":"string"},"positionTitle":{"type":"string"},"dateApplied":{"type":"number"},"url":{"type":"string"},"compensation":{"type":"string"},"company":{"$ref":"#/components/schemas/CompanyViewModel"},"comments":{"type":"array","items":{"$ref":"#/components/schemas/CommentViewModel"}},"links":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationLinkViewModel"}}},"required":["status","locationType","id","site","positionTitle","dateApplied","url","compensation","company","comments"]},"ApplicationListViewModel":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationViewModel"}},"total":{"type":"number"}},"required":["data"]},"ApplicationsUploadViewModel":{"type":"object","properties":{"addHeaders":{"type":"string","format":"boolean"},"file":{"type":"string","format":"binary"}},"required":["addHeaders","file"]},"CompanyListViewModel":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/CompanyViewModel"}},"total":{"type":"number"}},"required":["data"]},"ApplicationNestedViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"status":{"allOf":[{"$ref":"#/components/schemas/EnumApplicationStatus"}]},"locationType":{"allOf":[{"$ref":"#/components/schemas/EnumLocationTypes"}]},"positionTitle":{"type":"string"},"dateApplied":{"type":"number"},"url":{"type":"string"},"compensation":{"type":"string"},"comments":{"type":"array","items":{"$ref":"#/components/schemas/CommentViewModel"}},"links":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationLinkViewModel"}},"id":{"type":"string"},"site":{"type":"string"}},"required":["status","locationType","positionTitle","dateApplied","url","compensation","comments","id","site"]},"CompanyFullViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"id":{"type":"string"},"name":{"type":"string"},"applications":{"type":"array","items":{"$ref":"#/components/schemas/ApplicationNestedViewModel"}}},"required":["id","name","applications"]},"CompanyFullListViewModel":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/CompanyFullViewModel"}},"total":{"type":"number"}},"required":["data"]},"EnumActivitySource":{"type":"number","enum":[-1,1],"x-enum-varnames":["None","Strava"]},"ActivityTypeViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"id":{"type":"string"},"activities":{"type":"array","items":{"$ref":"#/components/schemas/ActivityViewModel"}},"name":{"type":"string"}},"required":["id","name"]},"EnumUnitTypes":{"type":"number","enum":[1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18],"x-enum-varnames":["Inches","Feet","Miles","Centimeters","Meters","Kilometers","Milliseconds","Seconds","Minutes","Hours","MetersPerSecond","KilometersPerHour","MilesPerHour","FeetPerSecond","String","Number","Boolean","Date"]},"AttributeTypeViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"feature":{"allOf":[{"$ref":"#/components/schemas/EnumFeatures"}]},"id":{"type":"string"},"name":{"type":"string"}},"required":["id","name"]},"ActivityAttributeViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"unit":{"allOf":[{"$ref":"#/components/schemas/EnumUnitTypes"}]},"unitDisplay":{"allOf":[{"$ref":"#/components/schemas/EnumUnitTypes"}]},"id":{"type":"string"},"activity":{"$ref":"#/components/schemas/ActivityViewModel"},"attributeType":{"$ref":"#/components/schemas/AttributeTypeViewModel"},"value":{"type":"string"},"valueDisplay":{"type":"string"}},"required":["id","attributeType","value"]},"ActivityViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"source":{"allOf":[{"$ref":"#/components/schemas/EnumActivitySource"}]},"id":{"type":"string"},"activityType":{"$ref":"#/components/schemas/ActivityTypeViewModel"},"attributes":{"type":"array","items":{"$ref":"#/components/schemas/ActivityAttributeViewModel"}},"title":{"type":"string"},"weight":{"type":"number"},"duration":{"type":"number"},"description":{"type":"string"},"sourceId":{"type":"string"},"dateOccurred":{"type":"number"},"calories":{"type":"number"},"weightLost":{"type":"number"}},"required":["id","activityType","attributes","title","dateOccurred"]},"ActivityListViewModel":{"type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/ActivityViewModel"}},"total":{"type":"number"}},"required":["data"]},"ExerciseActivityUpload":{"type":"object","properties":{"file":{"type":"string","format":"binary"}},"required":["file"]},"ActivityTypeCreateViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"name":{"type":"string"}},"required":["name"]},"ActivityCreateViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"source":{"allOf":[{"$ref":"#/components/schemas/EnumActivitySource"}]},"title":{"type":"string"},"weight":{"type":"number"},"duration":{"type":"number"},"description":{"type":"string"},"sourceId":{"type":"string"},"dateOccurred":{"type":"number"},"activityType":{"$ref":"#/components/schemas/ActivityTypeCreateViewModel"},"attributes":{"type":"array","items":{"$ref":"#/components/schemas/ActivityAttributeCreateViewModel"}},"calories":{"type":"number"},"weightLost":{"type":"number"}},"required":["title","dateOccurred","activityType","attributes"]},"AttributeTypeCreateViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"feature":{"allOf":[{"$ref":"#/components/schemas/EnumFeatures"}]},"name":{"type":"string"}},"required":["name"]},"ActivityAttributeCreateViewModel":{"type":"object","properties":{"userId":{"type":"string"},"dateCreated":{"type":"number"},"dateUpdated":{"type":"number"},"unit":{"allOf":[{"$ref":"#/components/schemas/EnumUnitTypes"}]},"unitDisplay":{"allOf":[{"$ref":"#/components/schemas/EnumUnitTypes"}]},"value":{"type":"string"},"valueDisplay":{"type":"string"},"activity":{"$ref":"#/components/schemas/ActivityCreateViewModel"},"attributeType":{"$ref":"#/components/schemas/AttributeTypeCreateViewModel"}},"required":["value","attributeType"]},"StravaTokenViewModel":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"expirationDate":{"type":"number"}},"required":["accessToken"]}}}}